extracted Sidebar component
