+++ src/connected/Workspace.tsx
@@ -24,45 +24,21 @@
  * SOFTWARE.
  */
 
-import { ReactElement, useEffect, useRef, useState } from "react";
+import React, { CSSProperties, useEffect, useRef, useState } from "react";
 import { connect } from "react-redux";
 import { default as OlMap } from "ol/Map";
 import { Theme } from "@mui/system";
 import { WithStyles } from "@mui/styles";
 import createStyles from "@mui/styles/createStyles";
 import withStyles from "@mui/styles/withStyles";
-import InfoIcon from "@mui/icons-material/Info";
-import StackedLineChartIcon from "@mui/icons-material/StackedLineChart";
-import ThreeDRotationIcon from "@mui/icons-material/ThreeDRotation";
 
 import { AppState } from "@/states/appState";
 import SplitPane from "@/components/SplitPane";
 import Viewer from "./Viewer";
-import TimeSeriesPanel from "./TimeSeriesPanel";
-import VolumePanel from "./VolumePanel";
-import InfoPanel from "./InfoPanel";
-import Tabs from "@mui/material/Tabs";
-import Box from "@mui/material/Box";
-import Tab from "@mui/material/Tab";
-import i18n from "@/i18n";
-import { makeStyles } from "@/util/styles";
-import { setSidebarPanelId } from "@/actions/controlActions";
-import { SidebarPanelId, sidebarPanelIds } from "@/states/controlState";
-
-const sidebarPanelIcons: Record<SidebarPanelId, ReactElement> = {
-  info: <InfoIcon fontSize="inherit" />,
-  charts: <StackedLineChartIcon fontSize="inherit" />,
-  volume: <ThreeDRotationIcon fontSize="inherit" />,
-};
-
-const sidebarPanelLabels: Record<SidebarPanelId, string> = {
-  info: "Info",
-  charts: "Charts",
-  volume: "Volume",
-};
+import Sidebar from "./Sidebar";
 
 // Adjust for debugging split pane style
-const mapExtraStyle: React.CSSProperties = { padding: 0 };
+const mapExtraStyle: CSSProperties = { padding: 0 };
 
 const styles = (_theme: Theme) =>
   createStyles({
@@ -104,45 +80,25 @@ const styles = (_theme: Theme) =>
     },
   });
 
-const sxes = makeStyles({
-  tabs: { minHeight: "34px" },
-  tab: {
-    padding: "5px 10px",
-    textTransform: "none",
-    fontWeight: "regular",
-    minHeight: "32px",
-  },
-});
-
 interface WorkspaceProps extends WithStyles<typeof styles> {
   sidebarOpen: boolean;
-  sidebarPanelId: SidebarPanelId;
-  setSidebarPanelId: (sidebarPanelId: SidebarPanelId) => void;
 }
 
 // noinspection JSUnusedLocalSymbols
 const mapStateToProps = (state: AppState) => {
   return {
     sidebarOpen: state.controlState.sidebarOpen,
-    sidebarPanelId: state.controlState.sidebarPanelId,
   };
 };
 
-const mapDispatchToProps = {
-  setSidebarPanelId,
-};
+const mapDispatchToProps = {};
 
 type Layout = "hor" | "ver";
 const getLayout = (): Layout => {
   return window.innerWidth / window.innerHeight >= 1 ? "hor" : "ver";
 };
 
-const _Workspace: React.FC<WorkspaceProps> = ({
-  classes,
-  sidebarOpen,
-  sidebarPanelId,
-  setSidebarPanelId,
-}) => {
+const _Workspace: React.FC<WorkspaceProps> = ({ classes, sidebarOpen }) => {
   const [map, setMap] = useState<OlMap | null>(null);
   const [layout, setLayout] = useState<Layout>(getLayout());
   const resizeObserver = useRef<ResizeObserver | null>(null);
@@ -186,34 +142,7 @@ const _Workspace: React.FC<WorkspaceProps> = ({
         child2ClassName={detailsPaneClassName}
       >
         <Viewer onMapRef={setMap} />
-        {/* TODO: make own Sidebar component */}
-        <Box sx={{ width: "100%" }}>
-          <Box sx={{ borderBottom: 1, borderColor: "divider" }}>
-            <Tabs
-              value={sidebarPanelId}
-              onChange={(_, value) => {
-                setSidebarPanelId(value as SidebarPanelId);
-              }}
-              variant="scrollable"
-              sx={sxes.tabs}
-            >
-              {sidebarPanelIds.map((panelId) => (
-                <Tab
-                  key={panelId}
-                  icon={sidebarPanelIcons[panelId]}
-                  iconPosition="start"
-                  sx={sxes.tab}
-                  disableRipple
-                  value={panelId}
-                  label={i18n.get(sidebarPanelLabels[panelId])}
-                />
-              ))}
-            </Tabs>
-          </Box>
-          {sidebarPanelId === "info" && <InfoPanel />}
-          {sidebarPanelId === "charts" && <TimeSeriesPanel />}
-          {sidebarPanelId === "volume" && <VolumePanel />}
-        </Box>
+        <Sidebar />
       </SplitPane>
     );
   } else {
